Remove Export::SheetMappings
Move everything into Export::SubmissionEntryRow. Export::SheetMappings
wasn't earning its keep and we can't use it in the generator anyway.
